fix: aligner la fiche évènement sur les styles du prototype
Comparaison des styles calculés entre la webapp et le proto, mobile et
desktop, plutôt qu'à l'œil.

Écart principal, en desktop : le proto pose le hero et la ligne de synthèse
en pleine largeur, au-dessus des deux colonnes ; seul le contenu en dessous
se partage entre la colonne de lecture et le bloc d'inscription. La fiche
mettait le hero dans la colonne de gauche, qui le réduisait à 540 px au lieu
de 992. Les largeurs sont désormais celles du proto : conteneur 1040,
colonnes 680 / 280, gouttière 32.

Le reste tenait à des valeurs qui avaient dérivé : titre du hero en graisse
normale au lieu de gras, titres de section à 20 px au lieu de 22, pastille
de timeline en ocre au lieu de corail, description à 14 px au lieu de 15.

« M'envoyer un rappel », « Ajouter à mon calendrier » et « Voir le site »
étaient gris alors que ce sont des actions : elles reprennent le
`.btn-secondary.small` du proto, factorisé dans un utilitaire
`.btn-outline-coral`. Le liseré du bloc mesures reprend le vert inclusion du
proto, et les cartes affichent le mois en toutes lettres.

+++ src/components/events/event-card.tsx
@@ -86,13 +86,15 @@ export function EventCard({
   const hasImage = hasUsableImageValue && failedImageSrc !== displayImage;
   const imageLoaded = loadedImageSrc === displayImage;
 
+  // Mois en toutes lettres, comme sur la carte du prototype : la ligne
+  // « ville · date » est tronquée si besoin, mais elle se lit d'abord.
   const formattedStartDate = formatDateValue(dateEvent, "fr-FR", {
     day: "numeric",
-    month: "short",
+    month: "long",
   });
   const formattedEndDate = formatDateValue(dateFin, "fr-FR", {
     day: "numeric",
-    month: "short",
+    month: "long",
   });
   const formattedDate =
     formattedStartDate && formattedEndDate && getDateKey(dateEvent) !== getDateKey(dateFin)
